refactored into library, implemented message types gen
