+++ js_parser/test_js_parser.sh
@@ -1,1 +1,1 @@
-node `dirname "$0"`/parser.js < testfiles/1698148651_source2.dem
+node `dirname "$0"`/example/example.js < testfiles/1698148651_source2.dem
